Refactor ParcelGenerator
