+++ CityGeneration.cs
@@ -52,147 +52,6 @@ namespace StrategyGame
         }
     }
 
-    // Original ParcelGenerator
-    public static class ParcelGenerator
-    {
-        private const double Epsilon = 1e-9;
-
-        private static bool IsBad(Coordinate c) =>
-            double.IsNaN(c.X) || double.IsNaN(c.Y) ||
-            double.IsInfinity(c.X) || double.IsInfinity(c.Y);
-
-        public static List<Parcel> GenerateParcels(CityDataModel model)
-        {
-            var parcels = new List<Parcel>();
-            // Ensure there's a road network and an urban area to work with.
-            if (model.RoadNetwork == null || !model.RoadNetwork.Any() || model.UrbanArea == null)
-                return parcels;
-
-            var gf = Nts.GeometryFactory.Default;
-
-            // 1. Get all valid road line segments as Geometry.
-            var roadLines = model.RoadNetwork
-                .Where(seg =>
-                    !IsBad(new Coordinate(seg.X1, seg.Y1)) &&
-                    !IsBad(new Coordinate(seg.X2, seg.Y2)) &&
-                    Math.Abs(seg.X1 - seg.X2) + Math.Abs(seg.Y1 - seg.Y2) > Epsilon)
-                .Select(seg => gf.CreateLineString(new[]
-                {
-                    new Coordinate(seg.X1, seg.Y1),
-                    new Coordinate(seg.X2, seg.Y2)
-                }))
-                .ToList<Nts.Geometry>();
-
-            // 2. Add the urban area boundary to the same collection.
-            if (model.UrbanArea.IsValid)
-            {
-                roadLines.Add(model.UrbanArea.Boundary);
-            }
-
-            // 3. Union ALL lines together at once. This correctly nodes the entire geometry set.
-            var nodedLines = UnaryUnionOp.Union(roadLines);
-
-            // 4. Polygonize the fully noded line network.
-            var polygonizer = new Polygonizer();
-            polygonizer.Add(nodedLines);
-            var rawPolys = polygonizer.GetPolygons();
-            Debug.WriteLine($"[ParcelGenerator] Polygonizer produced {rawPolys.Count} raw polygons");
-
-            var blocks = rawPolys.OfType<Nts.Polygon>()
-                .Where(p => p.IsValid && p.Area > 1e-9)
-                .ToList();
-
-            // The rest of the process can now proceed.
-            model.RawBlocks = blocks;
-            parcels = GenerateParcelsFromBlocks(blocks);
-
-            Debug.WriteLine($"[ParcelGenerator] Final parcel count: {parcels.Count}");
-            return parcels;
-        }
-
-        public static List<Parcel> GenerateParcelsFromBlocks(List<Nts.Polygon> blocks)
-        {
-            var parcels = new List<Parcel>();
-            foreach (var block in blocks)
-            {
-                if (block.IsValid && !block.IsEmpty && block.Area > 1e-9)
-                    SubdividePolygon(block, parcels);
-            }
-            return parcels;
-        }
-
-        private static void SubdividePolygon(Nts.Polygon poly, List<Parcel> output)
-        {
-            const double MinParcelArea = 0.00005;
-            RecursiveSplit(poly, output, MinParcelArea);
-        }
-
-        private static void RecursiveSplit(Nts.Polygon poly, List<Parcel> output, double minArea)
-        {
-            // --- Start Diagnostic Logging ---
-            Debug.WriteLine($"Splitting polygon. Area: {poly.Area}, IsValid: {poly.IsValid}, Envelope: {poly.EnvelopeInternal}");
-            // --- End Diagnostic Logging ---
-
-            if (poly.Area < minArea * 1.5)
-            {
-                if (poly.IsValid && !poly.IsEmpty)
-                    output.Add(new Parcel { Shape = poly });
-                return;
-            }
-
-            var envelope = poly.EnvelopeInternal;
-            var gf = poly.Factory;
-            bool splitVertical = envelope.Width > envelope.Height;
-
-            Nts.Geometry splitLine;
-            if (splitVertical)
-            {
-                double midX = envelope.MinX + envelope.Width / 2;
-                splitLine = gf.CreateLineString(new[]
-                {
-                    new Nts.Coordinate(midX, envelope.MinY),
-                    new Nts.Coordinate(midX, envelope.MaxY)
-                });
-            }
-            else
-            {
-                double midY = envelope.MinY + envelope.Height / 2;
-                splitLine = gf.CreateLineString(new[]
-                {
-                    new Nts.Coordinate(envelope.MinX, midY),
-                    new Nts.Coordinate(envelope.MaxX, midY)
-                });
-            }
-
-            try
-            {
-                var splitGeometries = poly.Difference(splitLine);
-
-                // If the split did not create two or more pieces, stop recursion to avoid an infinite loop.
-                if (splitGeometries.NumGeometries < 2)
-                {
-                    if (poly.IsValid && !poly.IsEmpty)
-                        output.Add(new Parcel { Shape = poly });
-                    return;
-                }
-
-                for (int i = 0; i < splitGeometries.NumGeometries; i++)
-                {
-                    if (splitGeometries.GetGeometryN(i) is Nts.Polygon splitPoly && splitPoly.IsValid && !splitPoly.IsEmpty)
-                    {
-                        RecursiveSplit(splitPoly, output, minArea);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                // This will tell us if the split itself throws a C# error
-                Debug.WriteLine($"[RECURSIVE SPLIT ERROR] {ex.Message}");
-                if (poly.IsValid && !poly.IsEmpty)
-                    output.Add(new Parcel { Shape = poly });
-            }
-        }
-    }
 
     // Basic developer agent classes used by LandUseSimulator
     public abstract class DeveloperAgent
